Split generic enum method type argument diagnostics
